+++ Nos/Views/CompactNoteView.swift
@@ -6,37 +6,68 @@ import Dependencies
 /// it is too long
 struct CompactNoteView: View {
     
-    @Environment(\.managedObjectContext) private var viewContext
+    /// The note whose content will be displayed
+    private let note: Event
+    
+    /// The maximum number of lines to show before truncating (if `showFullMessage` is false)
+    private let truncationLineLimit = 12
+    
+    /// If true this view will truncate long notes and show a "Read more" button to view the whole thing. If false 
+    /// the full note will always be displayed
+    private var shouldTruncate: Bool
+    
+    /// Whether link previews should be displayed for links found in the note text.
+    private var showLinkPreviews: Bool
+    
+    /// If true links will be highlighted and open when tapped. If false the text will change to a secondary color
+    /// and links will not be tappable.
+    private var allowUserInteraction: Bool
+    
+    /// Whether this view is currently displayed in a truncated state
+    @State private var isTextTruncated = true
 
-    var note: Event
-
-    @State private var shouldShowReadMore = false
-    @State var showFullMessage: Bool
+    /// The size of the full note text 
     @State private var intrinsicSize = CGSize.zero
+    
+    /// The size of the note text truncated to `truncationLineLimit` lines.
     @State private var truncatedSize = CGSize.zero
-    private var loadLinks: Bool
     
+    @Environment(\.managedObjectContext) private var viewContext
     @EnvironmentObject private var router: Router
-    @Dependency(\.persistenceController) private var persistenceController
     
-    internal init(note: Event, showFullMessage: Bool = false, loadLinks: Bool = true) {
-        _showFullMessage = .init(initialValue: showFullMessage)
+    internal init(
+        note: Event, 
+        shouldTruncate: Bool = false, 
+        showLinkPreviews: Bool = true,
+        allowUserInteraction: Bool = true
+    ) {
         self.note = note
-        self.loadLinks = loadLinks
+        self.shouldTruncate = shouldTruncate
+        self.showLinkPreviews = showLinkPreviews
+        self.allowUserInteraction = allowUserInteraction
     }
     
-    func updateShouldShowReadMore() {
-        shouldShowReadMore = intrinsicSize.height > truncatedSize.height + 30
+    /// Calculates whether the note text is long enough to need truncation given `truncationLineLimit`.
+    var noteNeedsTruncation: Bool {
+        shouldTruncate && intrinsicSize.height > truncatedSize.height + 30 
+    }
+    
+    /// Calculates whether the Read More button should be shown. 
+    var showReadMoreButton: Bool {
+        noteNeedsTruncation && isTextTruncated
     }
     
     var formattedText: some View {
         noteText
             .font(.body)
-            .foregroundColor(.primaryTxt)
-            .tint(.accent) 
+            .foregroundColor(allowUserInteraction ? .primaryTxt : .secondaryTxt)
+            .tint(allowUserInteraction ? .accent : .secondaryTxt) 
             .padding(15)
             .textSelection(.enabled)
             .environment(\.openURL, OpenURLAction { url in
+                guard allowUserInteraction else {
+                    return .handled
+                }
                 router.open(url: url, with: viewContext)
                 return .handled
             })
@@ -57,12 +88,12 @@ struct CompactNoteView: View {
 
     var body: some View {
         VStack(alignment: .leading, spacing: 0) {
-            if showFullMessage {
+            if !isTextTruncated || !shouldTruncate {
                 formattedText
                     .fixedSize(horizontal: false, vertical: true)
             } else {
                 formattedText
-                    .lineLimit(12)
+                    .lineLimit(truncationLineLimit)
                     .background {
                         GeometryReader { geometryProxy in
                             Color.clear.preference(key: TruncatedSizePreferenceKey.self, value: geometryProxy.size)
@@ -71,11 +102,13 @@ struct CompactNoteView: View {
                     .onPreferenceChange(TruncatedSizePreferenceKey.self) { newSize in
                         if newSize.height > truncatedSize.height {
                             truncatedSize = newSize
-                            updateShouldShowReadMore()
                         }
                     }
                     .background {
-                        noteText
+                        // To calculate whether the text should be truncated we create this hidden view of 
+                        // `formattedText` and record its size. It is compared to `truncatedSize` and used in the 
+                        // calculation of `noteNeedsTruncation`.
+                        formattedText
                             .fixedSize(horizontal: false, vertical: true)
                             .hidden()
                             .background {
@@ -89,16 +122,15 @@ struct CompactNoteView: View {
                             .onPreferenceChange(IntrinsicSizePreferenceKey.self) { newSize in
                                 if newSize.height > intrinsicSize.height {
                                     intrinsicSize = newSize
-                                    updateShouldShowReadMore()
                                 }
                             }
                     }
             }
-            if shouldShowReadMore && !showFullMessage {
+            if showReadMoreButton {
                 ZStack(alignment: .center) {
                     Button {
                         withAnimation {
-                            showFullMessage = true
+                            isTextTruncated = false
                         }
                     } label: {
                         Text(String(localized: .localizable.readMore).uppercased())
@@ -112,13 +144,13 @@ struct CompactNoteView: View {
                 .frame(maxWidth: .infinity)
                 .padding(EdgeInsets(top: 0, leading: 0, bottom: 10, trailing: 0))
             }
-            if note.kind == EventKind.text.rawValue, loadLinks, !note.contentLinks.isEmpty {
+            if note.kind == EventKind.text.rawValue, showLinkPreviews, !note.contentLinks.isEmpty {
                 LinkPreviewCarousel(links: note.contentLinks)
             }
         }
         .frame(maxWidth: .infinity, alignment: .leading)
-        .onChange(of: note.attributedContent) {
-            updateShouldShowReadMore()
+        .task {
+            await note.loadViewData()
         }
     }
 }
@@ -139,12 +171,11 @@ struct CompactNoteView_Previews: PreviewProvider {
     
     static var previews: some View {
         Group {
-            CompactNoteView(note: previewData.linkNote)
+            CompactNoteView(note: previewData.linkNote, allowUserInteraction: false)
             CompactNoteView(note: previewData.shortNote)
             CompactNoteView(note: previewData.longNote)
-            CompactNoteView(note: previewData.longFormNote)
             CompactNoteView(note: previewData.doubleImageNote)
-            CompactNoteView(note: previewData.doubleImageNote, loadLinks: false)
+            CompactNoteView(note: previewData.doubleImageNote, showLinkPreviews: false)
         }
         .padding()
         .background(Color.previewBg)