+++ Nos/Views/Form/NosTextField.swift
@@ -4,12 +4,10 @@ import SwiftUINavigation
 struct NosTextField: View {
     
     var label: LocalizedStringResource
-    var footnote: LocalizedStringResource?
     @Binding var text: String
-    @FocusState private var focus: Bool
     
     var body: some View {
-        NosFormField(label: label, footnote: footnote) { 
+        NosFormField(label: label) {
             TextField("", text: $text)
                 .textInputAutocapitalization(.none)
                 .foregroundColor(.primaryTxt)