+++ Nos/Views/Home/HomeFeedView.swift
@@ -189,13 +189,6 @@ struct HomeFeedView: View {
     }
 }
 
-private let itemFormatter: DateFormatter = {
-    let formatter = DateFormatter()
-    formatter.dateStyle = .short
-    formatter.timeStyle = .medium
-    return formatter
-}()
-
 struct ContentView_Previews: PreviewProvider {
     
     static var previewData = PreviewData()