+++ Nos/Views/MutesView.swift
@@ -3,10 +3,6 @@ import SwiftUI
 struct MutesDestination: Hashable { }
 
 struct MutesView: View {
-    @Environment(\.managedObjectContext) private var viewContext
-    @EnvironmentObject private var relayService: RelayService
-    @EnvironmentObject private var router: Router
-
     @FetchRequest
     private var authors: FetchedResults<Author>