+++ CHANGELOG.md
@@ -8,10 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added a new UI for replying to messages that allows attaching images and setting an expiration date.
+
+## [0.1.7] - 2024-03-21Z
+
 - Fix issue in Profile title bar that displayed NIP-05 incorrectly.
 - Update font styles on Thread, Edit Profile, and Settings screens.
 - Fix issue with uploading photos on Mac.
 - Re-design the confirmation dialog that appears when you delete your NIP-05.
+- Fixed a bug where liking a note could cause other notes to appear liked.
+- Added a new UI for replying to messages that allows attaching images and setting an expiration date.
+- Fixed an issue where the "Read more" button could display on notes that didn't need it.
 
 ## [0.1.6] - 2024-03-07Z